Add in a Context helper
I hate having to reimplement this pattern, let's just stuff it in here
for now. This will at least allow for consistent and central context
handling accross the suite.

Also move to logrus. Anything user-facing ought to be logrus, but I'm
still reluctant to move client libraries in. Maybe I'll swap it for an
interface at some point.